fix the bug of bound inference from estimates

+++ apps/scheduling-tool/test/10.cpp
@@ -19,10 +19,10 @@ int main(int argc, char **argv) {
     Func f_u8("f_u8");
     Func f_u64_1("f_u64_1");
     Func f_u64_2("f_u64_2");
-    Buffer<uint8_t> a(1024 * 1024 + 2);
+    Buffer<float> a(1024 * 1024 + 2);
 
     f_u8(x) = (min(a(x) + 1, 17) * a(x+1) + a(x+2)) * a(x) * a(x) * a(x + 1) * a(x + 1);
-    f_u64_1(x) = cast<uint64_t>(f_u8(x)) + 1;
+    f_u64_1(x) = cast<float>(f_u8(x)) + 1;
     f_u64_2(x) = f_u64_1(x) * 3;
 
     // Ignoring the types, it would make sense to inline
